Separated metadata handling from PillarRotationTrait
